Implement working operator and equals buttons
When an operator button is pressed, such as +, we
first check if the first number has already been
stored. If not, we store the number on the display
and the operator selected, then we clear the
display. If it has already been stored, we store
second number and operate on it and display the
result. And the equals button does the same as if
the first number is stored and the display is not
empty.